# docs/archive/old-files/sourcecode/الخريطه الذهنيه/crowdshipping_project_complete/openapi_crowdshipping_expanded.yaml
openapi: 3.0.3
info:
  title: Crowdshipping API (Expanded)
  description: |
    Bilingual API spec for the Crowdshipping platform — includes auth, traveler-location,
    recommendations, products, auctions, bids, payments, and webhooks.
    مواصفات واجهات برمجة التطبيقات لمنصة Crowdshipping — تشمل المصادقة، موقع المسافر، التوصيات، المنتجات، المزادات، العروض، الدفع، وwebhooks.
  version: "1.0.0"
servers:
  - url: https://api.crowdshipping.example.com/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Local (development)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        seller_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        category:
          type: string
        origin_country:
          type: string
        price:
          type: number
        listing_type:
          type: string
          enum: [buy_now, auction, make_offer]
        location_hint:
          type: object
          properties:
            airport:
              type: string
            city:
              type: string
            lat:
              type: number
            lon:
              type: number
    Auction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
        seller_id:
          type: string
        start_price:
          type: number
        current_price:
          type: number
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, live, ended, canceled]
...
paths:
...
  /products:
    post:
      summary: Create a product listing — إنشاء منتج
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Created
    get:
      summary: Search products — البحث عن منتجات
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: origin_country
          in: query
          schema:
            type: string
        - name: listing_type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Results
...
  /auctions:
    post:
      summary: Create auction — إنشاء مزاد
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Auction'
      responses:
        '201':
          description: Auction created
...
security:
  - BearerAuth: []